Refactored login_page.dart textfields
